Extract loop detection into DependencyAnalyzer.ts
